Reuse existing assertions and tests for msgspec inheritance
