+++ tests/main/jsonschema/test_main_jsonschema.py
@@ -87,6 +87,7 @@ from tests.main.conftest import (
     MSGSPEC_LEGACY_BLACK_SKIP,
     TIMESTAMP,
     _assert_model_json_invalid,
+    _assert_python_module_importable,
     _generated_model,
     _generated_package_module,
     _model_json_validator,
@@ -19786,3 +19787,202 @@ def test_msgspec_enum_diagnostics_preserve_integer_comparison_error(output_file:
         output_model_type=DataModelType.MsgspecStruct,
     )
     assert_output(f"{output_file.exists()}\n", EXPECTED_JSON_SCHEMA_PATH / "msgspec_enum_diagnostics/absent.txt")
+
+
+MSGSPEC_INHERITANCE_DATA = DATA_PATH / "msgspec_inheritance"
+
+
+MSGSPEC_INHERITANCE_EXPECTED = EXPECTED_MAIN_PATH / "msgspec_inheritance"
+
+
+MSGSPEC_INHERITANCE_CASES = json.loads((MSGSPEC_INHERITANCE_DATA / "cases.json").read_text())
+
+
+@pytest.mark.parametrize("entrypoint", ["cli", "api"])
+@pytest.mark.parametrize("case", [case for case in MSGSPEC_INHERITANCE_CASES if case["error"]], ids=itemgetter("name"))
+def test_msgspec_inheritance_reject_conflicting_layouts(
+    output_file: Path, capsys: pytest.CaptureFixture[str], entrypoint: str, case: dict
+) -> None:
+    """Diagnose actual slot conflicts without writing unusable generated modules."""
+    expected = MSGSPEC_INHERITANCE_EXPECTED / f"{case['name']}.txt"
+    if entrypoint == "cli":
+        run_main_and_assert(
+            input_path=MSGSPEC_INHERITANCE_DATA / case["schema"],
+            output_path=output_file,
+            input_file_type=case["format"],
+            extra_args=["--output-model-type", "msgspec.Struct", *case["cli"]],
+            expected_exit=Exit.ERROR,
+            capsys=capsys,
+            expected_stderr=expected.read_text(),
+            output_should_not_exist=True,
+        )
+    else:
+        with pytest.raises(Error) as error:
+            generate(
+                MSGSPEC_INHERITANCE_DATA / case["schema"],
+                output=output_file,
+                input_file_type=InputFileType(case["format"]),
+                output_model_type=DataModelType.MsgspecStruct,
+                **case["options"],
+            )
+        assert_output(f"{error.value}\n", expected)
+        assert_output(f"{output_file.exists()}\n", MSGSPEC_INHERITANCE_EXPECTED / "absent.txt")
+
+
+@pytest.mark.parametrize("entrypoint", ["cli", "api"])
+@pytest.mark.parametrize(
+    "case", [case for case in MSGSPEC_INHERITANCE_CASES if not case["error"]], ids=itemgetter("name")
+)
+def test_msgspec_inheritance_preserve_compatible_layouts(output_file: Path, entrypoint: str, case: dict) -> None:
+    """Preserve generated bytes, inherited slots, base order, MRO, and real decoding."""
+    import msgspec
+
+    options = case["options"].copy()
+    if "custom_template_dir" in options:
+        options["custom_template_dir"] = MSGSPEC_INHERITANCE_DATA / "templates"
+
+    if entrypoint == "cli":
+        run_main_and_assert(
+            input_path=MSGSPEC_INHERITANCE_DATA / case["schema"],
+            output_path=output_file,
+            input_file_type=case["format"],
+            extra_args=[
+                "--output-model-type",
+                "msgspec.Struct",
+                "--target-python-version",
+                "3.10",
+                "--disable-timestamp",
+                *case["cli"],
+            ],
+            assert_func=assert_file_content,
+            expected_file=MSGSPEC_INHERITANCE_EXPECTED / f"{case['name']}.py",
+            force_exec_validation=True,
+        )
+    else:
+        run_generate_file_and_assert(
+            input_path=MSGSPEC_INHERITANCE_DATA / case["schema"],
+            output_path=output_file,
+            input_file_type=InputFileType(case["format"]),
+            output_model_type=DataModelType.MsgspecStruct,
+            target_python_version=PythonVersion.PY_310,
+            disable_timestamp=True,
+            assert_func=assert_file_content,
+            expected_file=MSGSPEC_INHERITANCE_EXPECTED / f"{case['name']}.py",
+            unchanged_inputs={"options": options},
+            **options,
+        )
+    with _generated_model(output_file, "generated_msgspec_inheritance", "Payload") as model:
+        converted = msgspec.convert(case["payload"], type=model)
+        decoded = msgspec.json.decode(json.dumps(case["payload"]), type=model)
+        result = {
+            "convert": msgspec.to_builtins(converted),
+            "decode": msgspec.to_builtins(decoded),
+            "fields": model.__struct_fields__,
+            "bases": [base.__name__ for base in model.__bases__],
+            "mro": [base.__name__ for base in model.__mro__],
+        }
+        with pytest.raises(msgspec.ValidationError):
+            msgspec.convert({}, type=model)
+        with pytest.raises(msgspec.ValidationError):
+            msgspec.json.decode(b"{}", type=model)
+    assert_output(json.dumps(result, indent=2) + "\n", MSGSPEC_INHERITANCE_EXPECTED / f"{case['name']}.txt")
+
+
+@pytest.mark.parametrize("entrypoint", ["cli", "api"])
+@pytest.mark.parametrize("input_format", ["jsonschema", "graphql"])
+@pytest.mark.parametrize(
+    "backend",
+    [
+        DataModelType.PydanticV2BaseModel,
+        DataModelType.PydanticV2Dataclass,
+        DataModelType.DataclassesDataclass,
+        DataModelType.TypingTypedDict,
+    ],
+)
+def test_msgspec_inheritance_preserve_other_backends(
+    output_file: Path, entrypoint: str, input_format: str, backend: DataModelType
+) -> None:
+    """Retain valid multiple generated bases for every other model backend."""
+    case = next(case for case in MSGSPEC_INHERITANCE_CASES if case["name"] == f"{input_format}_independent")
+    expected = MSGSPEC_INHERITANCE_EXPECTED / f"other_{input_format}_{backend.value.replace('.', '_')}.py"
+    if entrypoint == "cli":
+        run_main_and_assert(
+            input_path=MSGSPEC_INHERITANCE_DATA / case["schema"],
+            output_path=output_file,
+            input_file_type=input_format,
+            extra_args=[
+                "--output-model-type",
+                backend.value,
+                "--target-python-version",
+                "3.10",
+                "--disable-timestamp",
+                *case["cli"],
+            ],
+            assert_func=assert_file_content,
+            expected_file=expected,
+            force_exec_validation=True,
+        )
+    else:
+        run_generate_file_and_assert(
+            input_path=MSGSPEC_INHERITANCE_DATA / case["schema"],
+            output_path=output_file,
+            input_file_type=InputFileType(input_format),
+            output_model_type=backend,
+            target_python_version=PythonVersion.PY_310,
+            disable_timestamp=True,
+            assert_func=assert_file_content,
+            expected_file=expected,
+            **case["options"],
+        )
+    if backend == DataModelType.TypingTypedDict:
+        import msgspec
+
+        with _generated_model(output_file, "generated_other_backend_inheritance", "Payload") as model:
+            converted = msgspec.convert(case["payload"], type=model)
+            decoded = msgspec.json.decode(json.dumps(case["payload"]), type=model)
+            assert_output(
+                json.dumps({"convert": converted, "decode": decoded}, sort_keys=True) + "\n",
+                DATA_PATH / "payloads/msgspec_inheritance_typed_dict.txt",
+            )
+            with pytest.raises(msgspec.ValidationError, match="missing required field"):
+                msgspec.convert({}, type=model)
+            with pytest.raises(msgspec.ValidationError, match="missing required field"):
+                msgspec.json.decode("{}", type=model)
+        return
+    assert_generated_model_json_validation(
+        output_file,
+        module_name="generated_other_backend_inheritance",
+        model_name="Payload",
+        valid_json=json.dumps(case["payload"]),
+        invalid_json="{}",
+        expected_error_type="missing",
+        expected_attribute_path=("c",),
+        expected_attribute_value=3,
+    )
+
+
+@pytest.mark.parametrize("entrypoint", ["cli", "api"])
+def test_msgspec_inheritance_preserve_mixed_opaque_layout(output_file: Path, entrypoint: str) -> None:
+    """Keep opaque ancestry conservative even after a known compact conflict."""
+    if entrypoint == "cli":
+        run_main_and_assert(
+            input_path=MSGSPEC_INHERITANCE_DATA / "dry_opaque.json",
+            output_path=output_file,
+            input_file_type="jsonschema",
+            extra_args=["--output-model-type", "msgspec.Struct", "--disable-timestamp"],
+            assert_func=assert_file_content,
+            expected_file=MSGSPEC_INHERITANCE_EXPECTED / "dry_opaque.py",
+        )
+    else:
+        run_generate_file_and_assert(
+            input_path=MSGSPEC_INHERITANCE_DATA / "dry_opaque.json",
+            output_path=output_file,
+            input_file_type=InputFileType.JsonSchema,
+            output_model_type=DataModelType.MsgspecStruct,
+            disable_timestamp=True,
+            assert_func=assert_file_content,
+            expected_file=MSGSPEC_INHERITANCE_EXPECTED / "dry_opaque.py",
+        )
+    with pytest.raises(TypeError) as error:
+        _assert_python_module_importable(output_file, "generated_mixed_opaque", "Payload")
+    assert_output(f"{error.value}\n", MSGSPEC_INHERITANCE_EXPECTED / "dry_opaque.txt")
