--- tests/main/test_msgspec_inheritance.py
@@ -1,218 +1,0 @@
-"""End-to-end diagnostics and compatible generated msgspec inheritance layouts."""
-
-from __future__ import annotations
-
-import json
-from operator import itemgetter
-from typing import TYPE_CHECKING
-
-import pytest
-
-from datamodel_code_generator import DataModelType, Error, InputFileType, PythonVersion, generate
-from datamodel_code_generator.__main__ import Exit
-from tests.conftest import assert_output, create_assert_file_content
-from tests.main.conftest import (
-    DATA_PATH,
-    EXPECTED_MAIN_PATH,
-    _assert_python_module_importable,
-    _generated_model,
-    assert_generated_model_json_validation,
-    run_generate_file_and_assert,
-    run_main_and_assert,
-)
-
-if TYPE_CHECKING:
-    from pathlib import Path
-
-DATA = DATA_PATH / "msgspec_inheritance"
-EXPECTED = EXPECTED_MAIN_PATH / "msgspec_inheritance"
-CASES = json.loads((DATA / "cases.json").read_text())
-assert_file_content = create_assert_file_content(EXPECTED)
-
-
-@pytest.mark.parametrize("entrypoint", ["cli", "api"])
-@pytest.mark.parametrize("case", [case for case in CASES if case["error"]], ids=itemgetter("name"))
-def test_msgspec_inheritance_reject_conflicting_layouts(
-    output_file: Path, capsys: pytest.CaptureFixture[str], entrypoint: str, case: dict
-) -> None:
-    """Diagnose actual slot conflicts without writing unusable generated modules."""
-    expected = EXPECTED / f"{case['name']}.txt"
-    if entrypoint == "cli":
-        run_main_and_assert(
-            input_path=DATA / case["schema"],
-            output_path=output_file,
-            input_file_type=case["format"],
-            extra_args=["--output-model-type", "msgspec.Struct", *case["cli"]],
-            expected_exit=Exit.ERROR,
-            capsys=capsys,
-            expected_stderr=expected.read_text(),
-            output_should_not_exist=True,
-        )
-    else:
-        with pytest.raises(Error) as error:
-            generate(
-                DATA / case["schema"],
-                output=output_file,
-                input_file_type=InputFileType(case["format"]),
-                output_model_type=DataModelType.MsgspecStruct,
-                **case["options"],
-            )
-        assert_output(f"{error.value}\n", expected)
-        assert_output(f"{output_file.exists()}\n", EXPECTED / "absent.txt")
-
-
-@pytest.mark.parametrize("entrypoint", ["cli", "api"])
-@pytest.mark.parametrize("case", [case for case in CASES if not case["error"]], ids=itemgetter("name"))
-def test_msgspec_inheritance_preserve_compatible_layouts(output_file: Path, entrypoint: str, case: dict) -> None:
-    """Preserve generated bytes, inherited slots, base order, MRO, and real decoding."""
-    import msgspec
-
-    options = case["options"].copy()
-    if "custom_template_dir" in options:
-        options["custom_template_dir"] = DATA / "templates"
-
-    if entrypoint == "cli":
-        run_main_and_assert(
-            input_path=DATA / case["schema"],
-            output_path=output_file,
-            input_file_type=case["format"],
-            extra_args=[
-                "--output-model-type",
-                "msgspec.Struct",
-                "--target-python-version",
-                "3.10",
-                "--disable-timestamp",
-                *case["cli"],
-            ],
-            assert_func=assert_file_content,
-            expected_file=f"{case['name']}.py",
-            force_exec_validation=True,
-        )
-    else:
-        run_generate_file_and_assert(
-            input_path=DATA / case["schema"],
-            output_path=output_file,
-            input_file_type=InputFileType(case["format"]),
-            output_model_type=DataModelType.MsgspecStruct,
-            target_python_version=PythonVersion.PY_310,
-            disable_timestamp=True,
-            assert_func=assert_file_content,
-            expected_file=f"{case['name']}.py",
-            unchanged_inputs={"options": options},
-            **options,
-        )
-    with _generated_model(output_file, "generated_msgspec_inheritance", "Payload") as model:
-        converted = msgspec.convert(case["payload"], type=model)
-        decoded = msgspec.json.decode(json.dumps(case["payload"]), type=model)
-        result = {
-            "convert": msgspec.to_builtins(converted),
-            "decode": msgspec.to_builtins(decoded),
-            "fields": model.__struct_fields__,
-            "bases": [base.__name__ for base in model.__bases__],
-            "mro": [base.__name__ for base in model.__mro__],
-        }
-        with pytest.raises(msgspec.ValidationError):
-            msgspec.convert({}, type=model)
-        with pytest.raises(msgspec.ValidationError):
-            msgspec.json.decode(b"{}", type=model)
-    assert_output(json.dumps(result, indent=2) + "\n", EXPECTED / f"{case['name']}.txt")
-
-
-@pytest.mark.parametrize("entrypoint", ["cli", "api"])
-@pytest.mark.parametrize("input_format", ["jsonschema", "graphql"])
-@pytest.mark.parametrize(
-    "backend",
-    [
-        DataModelType.PydanticV2BaseModel,
-        DataModelType.PydanticV2Dataclass,
-        DataModelType.DataclassesDataclass,
-        DataModelType.TypingTypedDict,
-    ],
-)
-def test_msgspec_inheritance_preserve_other_backends(
-    output_file: Path, entrypoint: str, input_format: str, backend: DataModelType
-) -> None:
-    """Retain valid multiple generated bases for every other model backend."""
-    case = next(case for case in CASES if case["name"] == f"{input_format}_independent")
-    expected = f"other_{input_format}_{backend.value.replace('.', '_')}.py"
-    if entrypoint == "cli":
-        run_main_and_assert(
-            input_path=DATA / case["schema"],
-            output_path=output_file,
-            input_file_type=input_format,
-            extra_args=[
-                "--output-model-type",
-                backend.value,
-                "--target-python-version",
-                "3.10",
-                "--disable-timestamp",
-                *case["cli"],
-            ],
-            assert_func=assert_file_content,
-            expected_file=expected,
-            force_exec_validation=True,
-        )
-    else:
-        run_generate_file_and_assert(
-            input_path=DATA / case["schema"],
-            output_path=output_file,
-            input_file_type=InputFileType(input_format),
-            output_model_type=backend,
-            target_python_version=PythonVersion.PY_310,
-            disable_timestamp=True,
-            assert_func=assert_file_content,
-            expected_file=expected,
-            **case["options"],
-        )
-    if backend == DataModelType.TypingTypedDict:
-        import msgspec
-
-        with _generated_model(output_file, "generated_other_backend_inheritance", "Payload") as model:
-            converted = msgspec.convert(case["payload"], type=model)
-            decoded = msgspec.json.decode(json.dumps(case["payload"]), type=model)
-            assert_output(
-                json.dumps({"convert": converted, "decode": decoded}, sort_keys=True) + "\n",
-                DATA_PATH / "payloads/msgspec_inheritance_typed_dict.txt",
-            )
-            with pytest.raises(msgspec.ValidationError, match="missing required field"):
-                msgspec.convert({}, type=model)
-            with pytest.raises(msgspec.ValidationError, match="missing required field"):
-                msgspec.json.decode("{}", type=model)
-        return
-    assert_generated_model_json_validation(
-        output_file,
-        module_name="generated_other_backend_inheritance",
-        model_name="Payload",
-        valid_json=json.dumps(case["payload"]),
-        invalid_json="{}",
-        expected_error_type="missing",
-        expected_attribute_path=("c",),
-        expected_attribute_value=3,
-    )
-
-
-@pytest.mark.parametrize("entrypoint", ["cli", "api"])
-def test_msgspec_inheritance_preserve_mixed_opaque_layout(output_file: Path, entrypoint: str) -> None:
-    """Keep opaque ancestry conservative even after a known compact conflict."""
-    if entrypoint == "cli":
-        run_main_and_assert(
-            input_path=DATA / "dry_opaque.json",
-            output_path=output_file,
-            input_file_type="jsonschema",
-            extra_args=["--output-model-type", "msgspec.Struct", "--disable-timestamp"],
-            assert_func=assert_file_content,
-            expected_file="dry_opaque.py",
-        )
-    else:
-        run_generate_file_and_assert(
-            input_path=DATA / "dry_opaque.json",
-            output_path=output_file,
-            input_file_type=InputFileType.JsonSchema,
-            output_model_type=DataModelType.MsgspecStruct,
-            disable_timestamp=True,
-            assert_func=assert_file_content,
-            expected_file="dry_opaque.py",
-        )
-    with pytest.raises(TypeError) as error:
-        _assert_python_module_importable(output_file, "generated_mixed_opaque", "Payload")
-    assert_output(f"{error.value}\n", EXPECTED / "dry_opaque.txt")
